# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TrapVal Identity API",
    "description": "Identity and authentication service for TrapVal.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": ".",
      "description": "Current Identity Service host"
    }
  ],
...
  "paths": {
...
    "/refresh-token": {
      "post": {
        "tags": ["Session"],
        "summary": "Refresh an access token",
        "description": "Rotates a refresh token for a device and returns a new token pair.",
        "operationId": "refreshToken",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefreshTokenRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/RefreshTokenSuccess" },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/oauth/sign-in-google": {
      "post": {
        "tags": ["OAuth"],
        "summary": "Sign in with Google",
        "description": "Verifies a Google ID token, creates the Google-backed account when needed, and returns an application token pair.",
        "operationId": "signInWithGoogle",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OAuthSignInRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/SignInSuccess" },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
...
    "/auth/find-email/{email}": {
      "get": {
        "tags": ["Password Recovery"],
        "summary": "Find a local account by email",
        "description": "Returns public local account data for the supplied email address.",
        "operationId": "findEmail",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "description": "Email address to look up.",
            "schema": { "type": "string", "format": "email" }
          }
        ],
        "responses": {
          "200": {
            "description": "Local account found.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponseEnvelope" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
...
    "/user/edit": {
      "put": {
        "tags": ["Profile"],
        "summary": "Edit the current user profile",
        "description": "Updates the supplied name and/or phone_number for the current user. Omitted fields remain unchanged. Through the API Gateway, provide a Bearer access token; the gateway validates its user_id claim and forwards X-User-ID. Direct Identity Service requests require the X-User-ID header (UUID).",
        "operationId": "editProfile",
        "security": [{ "BearerAuth": [] }, { "UserIDHeader": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EditProfileRequest" },
              "example": { "name": "Hafid", "phone_number": "081234567890" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" },
                "example": { "status": "success", "message": "Edit profile succesfuly" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token returned by a successful sign-in or sign-up. The API Gateway validates signature, expiration, and the user_id claim, then forwards X-User-ID to Identity Service."
      },
      "UserIDHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-User-ID",
        "description": "User UUID forwarded by the API Gateway after JWT validation. Used for direct Identity Service requests; the profile handlers do not independently authenticate this header."
      }
    },
    "responses": {
      "SignInSuccess": {
        "description": "Sign-in succeeded.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/AuthResponseEnvelope" },
            "example": {
              "status": "success",
              "message": "sign in success",
              "data": {
                "access_token": "eyJhbGciOiJIUzI1NiIs...",
                "refresh_token": "opaque-refresh-token",
                "token_type": "Bearer",
                "expires_in": 1800
              }
            }
          }
        }
      },
...
      "RefreshTokenSuccess": {
        "description": "Token refresh succeeded.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/AuthResponseEnvelope" },
            "example": {
              "status": "success",
              "message": "refresh token success",
              "data": {
                "access_token": "eyJhbGciOiJIUzI1NiIs...",
                "refresh_token": "new-opaque-refresh-token",
                "token_type": "Bearer",
                "expires_in": 1800
              }
            }
          }
        }
      },
      "BadRequest": {
        "description": "The payload, credentials, token, or requested operation is invalid.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "status": "error", "message": "payload validation failed" }
          }
        }
      },
...
    },
    "schemas": {
      "MessageResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["status", "message"],
        "properties": {
          "status": { "type": "string", "const": "success" },
          "message": { "type": "string" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["status", "message"],
        "properties": {
          "status": { "type": "string", "const": "error" },
          "message": { "type": "string" }
        }
      },
...
      "AuthResponseEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["status", "message", "data"],
        "properties": {
          "status": { "type": "string", "const": "success" },
          "message": { "type": "string" },
          "data": { "$ref": "#/components/schemas/AuthResponse" }
        }
      },
      "AuthResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["access_token", "refresh_token", "token_type", "expires_in"],
        "properties": {
          "access_token": { "type": "string", "description": "JWT access token." },
          "refresh_token": { "type": "string", "description": "Opaque refresh token." },
          "token_type": { "type": "string", "const": "Bearer" },
          "expires_in": { "type": "integer", "format": "int64", "const": 1800 }
        }
      },
...
      "UserResponseEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["status", "message", "data"],
        "properties": {
          "status": { "type": "string", "const": "success" },
          "message": { "type": "string", "const": "email found" },
          "data": { "$ref": "#/components/schemas/User" }
        }
      },
...
      "EditProfileRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": { "type": ["string", "null"], "minLength": 3, "maxLength": 128, "description": "Optional name. Omit or send null to leave unchanged." },
          "phone_number": { "type": ["string", "null"], "minLength": 3, "maxLength": 15, "description": "Optional phone number. Omit or send null to leave unchanged." }
        }
      },
      "User": {
        "type": "object",
        "description": "Public user data returned by email lookup and current profile endpoints.",
        "additionalProperties": false,
        "required": ["id", "username", "email", "phone_number", "provider", "provider_id", "avatar_url"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "username": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "phone_number": { "type": "string" },
          "provider": { "type": "string" },
          "provider_id": { "type": "string" },
          "avatar_url": { "type": "string" }
        }
      },
      "OAuthSignInRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id_token", "device_id"],
        "properties": {
          "id_token": { "type": "string", "description": "Google ID token." },
          "device_id": { "type": "string" }
        }
      },
...
      "RefreshTokenRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["refresh_token", "device_id"],
        "properties": {
          "refresh_token": { "type": "string" },
          "device_id": { "type": "string" }
        }
      },
...
    }
  }
}